Change activation policy for FORWARD chain TO DROP
Close the gap where a "router all accept" will allow connections established
in the setup window to continue afterwards by preventing routed traffic
during the activation.

By leaving the policy intact for INPUT and OUTPUT, there is no chance of
interfering with a remote connection being used to update the firewall.

Existing connections will have packets dropped during the activation period
but TCP connections should recover once the setup is finished.

Does nothing to address "client all accept" on a host, which should not
really be used on an internet-facing server unless you change the other
policies in the configuration as a minimum.